タイマー、カウンターのコンポーザブルの共通部分の切り出し、記述の統一
